deleted extra 001 file, addded jwtdown-fastapi and python-jose to req txt, added auth to backend api

+++ blog_service/routers/blogs_r.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter, Depends, Response, HTTPException, status
+from auth import authenticator
 from typing import Union, List, Optional
 from queries.blogs_q import (
     BlogError,
@@ -7,28 +8,20 @@ from queries.blogs_q import (
     BlogOut
     # BlogList
 )
-from token_auth import get_current_user
 
 router = APIRouter()
-
-not_authorized = HTTPException(
-    status_code=status.HTTP_401_UNAUTHORIZED,
-    detail="Invalid authentication credentials",
-    headers={"WWW-Authenticate": "Bearer"},
-)
-
 
 @router.post("/blogs", response_model=Union[BlogOut, BlogError])
 def create_blog(
     blog: BlogIn,
     response: Response,
     repo: BlogRepo = Depends(),
-    account: dict = Depends(get_current_user),
+    account: dict = Depends(authenticator.get_current_account_data),
 ):
-    print(account)
-    if "individual" not in account.user_type:
-        raise not_authorized
-    return repo.create(blog)
+    if account["user_type"] == 'individual':
+        return repo.create(blog)
+    else:
+        raise HTTPException(status_code=401, detail="Only registered users are allowed to post new blog")
 
 
 @router.get("/blogs", response_model=Union[BlogError, List[BlogOut]])
@@ -39,8 +32,15 @@ def all_blogs(
     return repo.all_blogs()
 
 @router.delete("/blogs/{blog_id}", response_model=bool)
-def delete_blog(blog_id: int, repo: BlogRepo = Depends()) -> bool:
-    return repo.delete(blog_id)
+def delete_blog(
+    blog_id: int,
+    repo: BlogRepo = Depends(),
+    account: dict = Depends(authenticator.get_current_account_data)
+    ) -> bool:
+    if account["user_type"] == 'individual':
+        return repo.delete(blog_id)
+    else:
+        raise HTTPException(status_code=401, detail="Only registered users are allowed to delete blog")
 
 
 @router.put("/blogs/{blog_id}", response_model=Union[BlogOut, BlogError])
@@ -48,9 +48,13 @@ def update_blog(
     blog_id: int,
     blog:BlogIn,
     repo: BlogRepo = Depends(),
+    account: dict = Depends(authenticator.get_current_account_data)
     ) ->Union[BlogOut, BlogError]:
+    if account["user_type"] == 'individual':
+        return repo.update(blog_id,blog)
+    else:
+        raise HTTPException(status_code=401, detail="Only registered users are allowed to update blog")
 
-    return repo.update(blog_id,blog)
 
 
 @router.get("/blogs/{blog_id}", response_model=Optional[BlogOut])
